ansible/requirements.txt
requests>=2.31.0,<3.0.0
PyYAML>=6.0,<7.0
cryptography==44.0.3
